requirements.txt
pandas
python-nmap